% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_calc_elevation_index.R
\name{di_calc_elevation_index}
\alias{di_calc_elevation_index}
\title{Calculate the elevation component of the defensibility index for a 30 m DEM}
\usage{
di_calc_elevation_index(dem)
}
\arguments{
\item{dem}{A 30 m DEM}
}
\value{
A raster of the elevation index
}
\description{
Calculate the elevation component of the defensibility index for a 30 m DEM
}
