% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_calc_visibility_index.R
\name{di_calc_visibility_index}
\alias{di_calc_visibility_index}
\title{Calculate the visibility component of the defensibility index for a 30 m DEM}
\usage{
di_calc_visibility_index(dem)
}
\arguments{
\item{dem}{A 30 m DEM}
}
\value{
A raster of the visibility index
}
\description{
Calculate the visibility component of the defensibility index for a 30 m DEM
}
