% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_calculate_defensibility_index.R
\name{di_calculate_defensibility_index}
\alias{di_calculate_defensibility_index}
\title{Calculate the defensibility index for a 30 meter DEM}
\usage{
di_calculate_defensibility_index(dem)
}
\arguments{
\item{dem}{A 30 m DEM}
}
\value{
A raster of the defensibility index
}
\description{
Calculate the defensibility index for a 30 meter DEM
}
