% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meve}
\alias{meve}
\title{a 30 meter raster of Mesa Verde National Park}
\format{
An object of class \code{RasterLayer} of dimension 727 x 643 x 1.
}
\usage{
meve
}
\description{
a 30 meter raster of Mesa Verde National Park
}
\keyword{datasets}
