% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.difference.angular.R
\name{elev.difference.angular}
\alias{elev.difference.angular}
\title{Calculate the angular elevation difference}
\usage{
elev.difference.angular(obs_elev, dem)
}
\arguments{
\item{obs_elev}{height of the observer, in meters}

\item{dem}{A 30 m DEM}
}
\value{
the angular elevation difference raster
}
\description{
NOTE: Build to only work on 30m DEMs!
}
